updated the 2023 tusket assesment script, plots, cleaned up counts, dealth with blueback proportions including modelling missing values.
Made a script for getting number of licences and catch for yarmouth shelburne 2023.

Fixed a bug in in the VR2W script.

modifed the one and two species escapement scripts to handle downstream migration. there used to be a seperate script for it, but now its baked into the original, as a T/F argument that swtiches on overwriting negative netcounts to 0.